requirements.txt
paho-mqtt
scipy
pygame
py-trees==0.8.3
networkx==2.2
xmlschema
numpy
shapely
imageio
ray
tqdm
numba
pandas
scikit-image
scikit-learn
opencv-python
h5py
wandb
pydot
decorator